Implement Bolding Adjacent Edges
In grid-component files, I changed how the grid is passing it's coordinates to the grid-cell component in order to let the cells be able to where they are in the grid. I also added an input to pass the grid to the grid-cells it contains.

In grid-cell-component.ts, I added functionality to bolding the adjacent edge by looking at the components grid and bolding the respective edges on keydown.