Fix Errors From No Grid on GridCell
In tests, there were errors because a gridCell on it's own does not have a grid tied to it so when we bold an edge, it throws an error because it does not have any adjacent cells to look at.

Fixed by adding if statements to check if there is a grid attached to the gridCell.